refactor(userspace-dp): extract afxdp/ha tunnel-purge; mod.rs is now a facade
Move the tunnel-remap purge ownership group out of afxdp/ha/mod.rs into a
dedicated `tunnel_purge` submodule (#6247), completing the split. Pure code
motion; tunnel-purge lock phasing is preserved byte-for-byte.

Relocated inherent methods:
- purge_remapped_tunnel_sessions (snapshot keys/deltas under the shared
  sessions lock, release, then delete + emit close deltas)
- push_purge_close_deltas (lossless close-delta push with dropped-frame
  accounting)

`tunnel_purge` references shared types only through fully-qualified
`crate::` paths, so it needs no module glob.

afxdp/ha/mod.rs is now a narrow facade: it declares the four owner
submodules (state, export, session_import, tunnel_purge), re-exports the
export handle types for afxdp/mod.rs and the server dispatcher, re-exports
the fair-drain helper under cfg(test), and includes the relocated
characterization tests. The retained `use super::*` re-exports the afxdp
namespace into `ha` so the `#[path]`-included ha_tests module resolves its
`use super::*` exactly as before (documented allow for the false-positive
unused-import lint).
